problem list and display fixed

+++ Item 3/Acme-Rookies/src/main/java/controllers/company/ProblemCompanyController.java
@@ -92,7 +92,7 @@ public class ProblemCompanyController extends AbstractController {
 	//List
 
 	@RequestMapping(value = "/list", method = RequestMethod.GET)
-	public ModelAndView list(@RequestParam final int positionId) {
+	public ModelAndView list() {
 		final ModelAndView res;
 		final Company company = this.companyService.findByPrincipal();
 		final Collection<Problem> problems = this.problemService.findProblemByCompany();
@@ -100,11 +100,12 @@ public class ProblemCompanyController extends AbstractController {
 		res = new ModelAndView("problem/list");
 		res.addObject("company", company);
 		res.addObject("problems", problems);
+		res.addObject("requestURI", "/problem/company/list");
 
 		return res;
 	}
-	// EDIT 
 
+	// EDIT 
 	@RequestMapping(value = "/edit", method = RequestMethod.GET)
 	public ModelAndView edit(@RequestParam final int problemId, @RequestParam final int positionId) {
 		ModelAndView result;
@@ -170,11 +171,14 @@ public class ProblemCompanyController extends AbstractController {
 
 		paramPositionId = request.getParameter("positionId");
 		final Integer positionId = paramPositionId.isEmpty() ? null : Integer.parseInt(paramPositionId);
-
-		final Problem problem = this.problemService.findOne(problemId);
-		this.problemService.delete(problem);
 		result = this.positionCompanyController.display(positionId);
-		result.addObject("problem", problem);
+		final Collection<Application> applications = this.applicationService.findAllByProblem(problemId);
+		if (applications.isEmpty()) {
+			final Problem problem = this.problemService.findOne(problemId);
+			this.problemService.delete(problem);
+			result.addObject("problem", problem);
+		} else
+			result.addObject("errortrace", "not.empty.applications");
 		return result;
 	}
